requirements.txt
Django==1.10.5
python-decouple==3.0
django-debug-toolbar==1.6
dj-database-url==0.4.2
django-compressor==2.1
django-compressor-toolkit==0.6.0
django-bower==5.2.0
djangorestframework==3.5.3
social-auth-app-django==0.1.0
django-crispy-forms==1.6.1